[Fix] 계산도우미 data class, adapter 수정

+++ app/src/main/java/com/example/moiz/data/network/dto/TravelDto.kt
@@ -209,7 +209,7 @@ data class BalancePercentDto(
     @SerializedName("nickname")
     val nickname:String?,
     @SerializedName("amount")
-    val amount:Int?,
+    val amount:Double?,
 )
 
 data class BillingHelperDto(
